Time: 83 ms (81.83%), Space: 198.7 MB (47.42%) - LeetHub

+++ 3453-separate-squares-i/3453-separate-squares-i.cpp
@@ -23,8 +23,8 @@ public:
             cut = false;
 
             for(auto& square: squares) {
-                const long bottom = square[1];
-                const long top = square[1] + square[2];
+                const int bottom = square[1];
+                const int top = square[1] + square[2];
                 if(bottom >= ans) {
                     upper += double(square[2])*square[2];
                 }
